src/auto_labelling_yolov8/requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0

# Deep Learning
ultralytics>=8.0.0  # YOLOv8
torch>=2.0.0
torchvision>=0.15.0

# API and HTTP
requests>=2.26.0
python-dotenv>=0.19.0

# Utilities
loguru>=0.6.0
pydantic>=2.0.0
python-jose>=3.3.0
python-multipart>=0.0.5

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Data processing
pandas==2.0.3
pycocotools==2.0.7
albumentations==1.3.1

# Utilities
tqdm==4.65.0
PyYAML==6.0.1
python-dotenv==1.0.0
loguru==0.7.2

# LLM Integration
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2

# Development
black==23.7.0
isort==5.12.0
flake8==6.1.0
mypy==1.5.1
pre-commit==3.3.3 